requirements.txt
Django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
PyMySQL==1.1.0
torch==2.7.0
torchvision==0.22.0
Pillow>=10.0.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0